aiohttp_server/requirements.txt
aiohttp
opentelemetry-instrumentation-digma==0.5.12
digma-instrumentation-testtools==0.5.12